usePlayer: extract persistence into usePlayerPersistence hook
Persistence was an 80-line useEffect inside usePlayer wiring up
three lifecycle hooks (beforeunload, pagehide, visibilitychange)
plus an interval to flush state to localStorage and the backend
backstop. Plus the related helpers (loadPersisted, the shape-
validating pickRestoreFromPersisted) and types (PersistedState,
the PERSIST_KEY constant). Self-contained: only reads stateRef.

Moved everything to `useHooks/usePlayerPersistence.ts`:
  - PERSIST_KEY + PERSIST_INTERVAL_MS (private to the module)
  - PersistedState (re-exported, since usePlayer's initializer
    consumes it)
  - loadPersisted() (re-exported)
  - pickRestoreFromPersisted() (re-exported)
  - usePlayerPersistence(stateRef) hook

usePlayer now has a single line: `usePlayerPersistence(stateRef)`.
The synchronous initializer still uses loadPersisted +
pickRestoreFromPersisted for the boot-time read; those just import
from the new file.

usePlayer.ts: 1216 → 1147 LOC (-69). usePlayerPersistence.ts is
178 lines including docstrings. tsc clean. vitest 36 passed.
